First commit of desktop app interacting with Wanikani API
This first stage only contains a command line tool to fetch Wanikani
data. Later we will implement the UI for it.

Wanikani API uses personal access tokens to authenticate to the API.
Therefore, we need to ask the user to provide their token and to store
it somewhere locally.

Wanikani has API guidelines that wish the developers to have local
caches for the requested data and use etag and last updated at to get
only delta data.

To achieve these two requirements, we use SQLite database to store the
access token and review statistics data locally. SQLite also supports
a handy feature of insert or replace which, ensures that we won't have
duplicates/old versions of data.

The statistics data is paged and we fetch all stats recursively using
the updated after feature.

Wanikani response types are built with common base object types with
varying data. We emulate this with generic types.