First version of working UI with async operations
Move source files under src directory to keep root clear.

Add ui flow that initializes db, trys to fetch token, and shows it. If
token is not found, ask it from the user, save it and show it.

Add a button to clear the token from db.

With token and no token views separated