Fetch user data from API to db and to app state
On user login now checks when each data set is last updated and then
fetches updates from API after that date. The data is inserted into
database and then all data is read to the elmish state.

Database.connection function is changed to be a variable and not a
function because there was some weird race condition with the null to
option mapping.